rewrote expansion port resource logic to work with a list. updated to i/o and expansion port structs in generic, dqbb, isepic, retroreplay, tpi, acia, expert, mmc64, warpspeed, reu, digimax, georam, magicvoice, midi, sfx sampler, sfx expander (all non-main, so cartsystem is sane again). updated c128 cart stuff, changed cartridge_get_file_name (see devlog), added monitor command "export" / "exp" that shows the expansion port devices, added snapshot hook to cart system, fixed reu and tfe disable bug, various other cleanups. breaking msvc with new files as usual :)
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@22623 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/vic20/cart/vic20cart.c
@@ -356,7 +356,7 @@ void cartridge_set_default(void)
     reset_try_flags();
 }
 
-const char *cartridge_get_file_name(WORD addr)
+const char *cartridge_get_file_name(int addr)
 {
     if (vic20cart_type == CARTRIDGE_VIC20_GENERIC) {
         /* special case handling for the multiple file generic type */
